Refactor logMessage.
- Use one logger named 'xicam'. In future any additional loggers should
be named 'xicam.'.
- Attach one handler to that top level logger. All message from logger
named 'xicam.*' will be emitted by that handler. This removes the need
for the blacklist, because we are not adding a handler to the root
    logger.
- Instead of hand-assembling a message string, we rely on a Formatter.
This lets us use threadName and timestamp that the logging framework is
already computing for us rather than doing again ourselves. Importantly,
this string munging is only performed if the message will actually be
emitted by at least one handler, so it will be possible to trade
performance and verbosity. (Currently this requires editing the source
to change the logger and handler minimum levels from "DEBUG" to
something else, but it could be made configurable in the future.)
- Note that we no longer let the caller pass in a custom logger name. If
they try, we ignore them and warn that we have ignored them. This
prevents accumualting an unbounded number of special loggers, which
cannot be garbage collected.
- We do however retain the smart inspection-based code by using the
logging framework's "extra" feature.